requirements.txt
beautifulsoup4>=4.9.3
discord.py>=2.0.0
lxml>=4.6.3
pytz>=2021.1
requests>=2.25.1
selenium>=4.0.0
